War Thunder victory!
By the grace of God, victory is had
with the making of the two War Thunder
commands! I just need to refine the
file location variables in fun.py and
pic.py when I am able to.

Thank you for reading, and I hope you
all have a grand day!